[REFAC]: TEST 코드 아키텍처 리팩토링
